# Optimise nil checking in rendering structs to Notifications.